\name{accIs}
\alias{accIs}
\title{Accuracy of FLR}
\description{
  Accuracy of the flr classification.
}
\usage{
  accIs(testData,testDataB)
}
\arguments{
  \item{testData}{an input data.frame of the test after classification.}
  \item{testDataB}{an input data.frame of the original test data.}
}
\value{
 return the accuracy of the classification
 }
 
 
 

