\name{dataset001}
\alias{dataset001}
\docType{data}
\title{
dataset001
}
\description{
Dataset with 296 instances if 25 attributes.
}
\usage{data(dataset001)}
\format{
  A data frame with 296 instances on the following 25 variables.
  \describe{
    \item{\code{state}}{9 US states.}
    \item{\code{county}}{County.}
	\item{\code{site.id}}{Site id.}
	\item{\code{latitude}}{Latitude.}
	\item{\code{longitude}}{Longtitude.}
    \item{\code{X2009.2011.dv}}{2009.2011 dv.}
    \item{\code{X2010.2012.dv}}{2010.2012 dv.}
    \item{\code{X2009.2011.design.value..ppm.2.3}}{2009-2011 design value (ppm)2,3}
	\item{\code{X2010.2012.design.value..ppm...estimated.}}{2010-2012 design value (ppm) [estimated].}
	\item{\code{X2009.2011.design.value.status4}}{2009-2011 design value status4.}
	\item{\code{percent.complete.in.20095}}{percent complete in 20095.}
	\item{\code{percent.complete.in.20105}}{percent complete in 20105.}
	\item{\code{percent.complete.in.20115}}{percent complete in 20115.}
	\item{\code{X2009.2011.average.percent.complete}}{2009-2011 average percent complete.}
	\item{\code{X..of.days.above.the.naaqs.in.2009}}{# of days above the naaqs in 2009.}
	\item{\code{X..of.days.above.the.naaqs.in.2010}}{# of days above the naaqs in 2010.}
	\item{\code{X..of.days.above.the.naaqs.in.2011}}{# of days above the naaqs in 2011.}
	\item{\code{X..of.days.above.the.naaqs.in.2012}}{# of days above the naaqs in 2012.}
	\item{\code{X4th.highest.daily.max.value.in.2009}}{4th highest daily max value in 2009.}
	\item{\code{X4th.highest.daily.max.value.in.2010}}{4th highest daily max value in 2010.}
	\item{\code{X4th.highest.daily.max.value.in.2011}}{4th highest daily max value in 2011.}
	\item{\code{X4th.highest.daily.max.value.in.2012.}}{4th highest daily max value in 2012.}
	\item{\code{column_27}}{Column_27.}
	\item{\code{column_29}}{Column_29.}
    \item{\code{class}}{Class category.}
  }
}
\source{
geocommons.com
}
\references{
geocommons.com
}
\keyword{datasets}

