\name{denormDatal}
\alias{denormDatal}
\title{Denormalize Fuzzy Lattices.}
\description{
  Denormalize fuzzy lattices.
}
\usage{
  denormDatal(fuzlat,bounds)
}
\arguments{
  \item{fuzlat}{a fuzzy lattice containing mix and max value for each instance of the data set at the first columns,from left to right, followed by className and categ.}
  \item{bounds}{a 2 comumn matrix containing min and max value for each instance of the dataset.}
}
\value{
 return denormalized fuzzy lattice.
 }
 
