\name{mat}
\alias{mat}
\docType{data}
\title{
Graph distance matrix
}
\description{
A matrix containing the distances of the nodes in a graph.}
\usage{data(mat)}
\format{
  A data frame of 9 rows and 9 columns.
  \describe{
    \item{\code{Illinois}}{number}
    \item{\code{Indiana}}{number}
    \item{\code{Kentucky}}{number}
    \item{\code{Michigan}}{number}
    \item{\code{North.Carolina}}{number}
    \item{\code{Ohio}}{number}
	\item{\code{Pennsylvania}}{number}
	\item{\code{Tennessee}}{number}
	\item{\code{Virginia}}{number}
  }
}
\keyword{datasets}
