\name{testD}
\alias{testD}
\alias{trainD}
\title{Creates Testing And Training Samples}
\description{
  Creates testing and training samples from the original data.
}
\usage{
  testD(data2)
  trainD(data2)
}
\arguments{
  \item{data2}{a data.frame flaged with the sepFlag function.}
}
\value{
 return the training and testing samples that will be used for the classification.
 }
 
 
 